+++ py_wake/rotor_avg_models/rotor_avg_model.py
@@ -1,5 +1,6 @@
 import numpy as np
 from numpy import newaxis as na
+import py_wake
 
 
 class RotorAvgModel():
@@ -72,7 +73,7 @@ class GridRotorAvg(RotorAvgModel):
 
 class EqGridRotorAvg(GridRotorAvg):
     def __init__(self, n):
-        X, Y = np.meshgrid(np.linspace(-1, 1, n + 2)[1:-1], np.linspace(-1, 1, n + 2)[1:-1])
+        X, Y = np.meshgrid(np.linspace(-1, 1, n + 2, dtype=py_wake.dtype)[1:-1], np.linspace(-1, 1, n + 2, dtype=py_wake.dtype)[1:-1])
         m = (X**2 + Y**2) < 1
         GridRotorAvg.__init__(self,
                               nodes_x=X[m].flatten(),
